+++ packages/app-features/core/hooks/use-workspaces.ts
@@ -1,6 +1,12 @@
 import { useCurrentUser } from './use-current-user'
 
-export const useGetTenants = () => {
+/**
+ * Get all workspaces of the current user
+ * Workspace is the same as organization
+ *
+ * @returns {Array} The workspaces
+ */
+export const useWorkspaces = () => {
   const currentUser = useCurrentUser()
 
   return (