+++ apps/desktop/CHANGELOG.md
@@ -1,5 +1,19 @@
 # electron-app
 
+## 0.14.1
+
+### Patch Changes
+
+- Updated dependencies [20103b7]
+- Updated dependencies [4178c80]
+- Updated dependencies [20103b7]
+- Updated dependencies [20103b7]
+- Updated dependencies [20103b7]
+  - @app/features@14.0.0
+  - @saas-ui-pro/feature-flags@0.12.1
+  - @saas-ui-pro/react@0.24.0
+  - @ui/theme@0.6.1
+
 ## 0.14.0
 
 ### Minor Changes